+++ bee-common/bee-common/src/shutdown_stream.rs
@@ -31,39 +31,41 @@ use futures::{
 
 use std::{marker::Unpin, pin::Pin};
 
+type Shutdown = oneshot::Receiver<()>;
+type FusedShutdown = future::Fuse<Shutdown>;
+
 /// A stream with a shutdown.
 ///
 /// This type wraps a shutdown receiver and a stream to produce a new stream that ends when the
 /// shutdown receiver is triggered or when the stream ends.
 pub struct ShutdownStream<S> {
-    shutdown: future::Fuse<oneshot::Receiver<()>>,
-    stream: S,
+    shutdown: FusedShutdown,
+    stream: stream::Fuse<S>,
 }
 
-impl<S: Stream> ShutdownStream<stream::Fuse<S>> {
+impl<S: Stream> ShutdownStream<S> {
     /// Create a new `ShutdownStream` from a shutdown receiver and an unfused stream.
     ///
     /// This method receives the stream to be wrapped and a `oneshot::Receiver` for the shutdown.
     /// Both the stream and the shutdown receiver are fused to avoid polling already completed
     /// futures.
-    pub fn new(shutdown: oneshot::Receiver<()>, stream: S) -> Self {
+    pub fn new(shutdown: Shutdown, stream: S) -> Self {
         Self {
             shutdown: shutdown.fuse(),
             stream: stream.fuse(),
         }
     }
-}
 
-impl<S: Stream + FusedStream> ShutdownStream<S> {
-    /// Create a new `ShutdownStream` from a shutdown receiver and a fused stream.
+    /// Create a new `ShutdownStream` from a fused shutdown receiver and a fused stream.
     ///
-    /// This method receives the fused stream to be wrapped and a `oneshot::Receiver` for the shutdown.
-    /// The shutdown receiver is fused to avoid polling already completed futures.
-    pub fn from_fused(shutdown: oneshot::Receiver<()>, stream: S) -> Self {
-        Self {
-            shutdown: shutdown.fuse(),
-            stream,
-        }
+    /// This method receives the fused stream to be wrapped and a fused `oneshot::Receiver` for the shutdown.
+    pub fn from_fused(shutdown: FusedShutdown, stream: stream::Fuse<S>) -> Self {
+        Self { shutdown, stream }
+    }
+
+    /// Consume and split the `ShutdownStream` into its shutdown receiver and stream.
+    pub fn split(self) -> (FusedShutdown, stream::Fuse<S>) {
+        (self.shutdown, self.stream)
     }
 }